+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/datagen/tag/ModEntityTypeTagsProvider.kt
@@ -18,17 +18,48 @@ class ModEntityTypeTagsProvider(
 ) : EntityTypeTagsProvider(pOutput, pProvider, GeneticsResequenced.ID, existingFileHelper) {
 
     companion object {
-        val SCRAPER_ENTITY_BLACKLIST: TagKey<EntityType<*>> = TagKey.create(
-            Registries.ENTITY_TYPE,
-            OtherUtil.modResource("scraper_blacklist")
-        )
+        private fun create(id: String): TagKey<EntityType<*>> {
+            return TagKey.create(Registries.ENTITY_TYPE, OtherUtil.modResource(id))
+        }
+
+        val SCRAPER_ENTITY_BLACKLIST = create("scraper_blacklist")
+        val AVOIDS_SCARE_CREEPER_GENE = create("avoids_scare_creeper_gene")
+        val AVOIDS_SCARE_ZOMBIE_GENE = create("avoids_scare_zombie_gene")
+        val AVOIDS_SCARE_SKELETON_GENE = create("avoids_scare_skeleton_gene")
+        val AVOIDS_SCARE_SPIDER_GENE = create("avoids_scare_spider_gene")
     }
 
     override fun addTags(pProvider: HolderLookup.Provider) {
-
         this.tag(SCRAPER_ENTITY_BLACKLIST)
             .add(EntityType.ARMOR_STAND, EntityType.PAINTING)
 
+        this.tag(AVOIDS_SCARE_CREEPER_GENE)
+            .add(EntityType.CREEPER)
+
+        this.tag(AVOIDS_SCARE_ZOMBIE_GENE)
+            .add(
+                EntityType.ZOMBIE,
+                EntityType.DROWNED,
+                EntityType.HUSK,
+                EntityType.ZOMBIE_VILLAGER,
+                EntityType.ZOMBIFIED_PIGLIN,
+                EntityType.ZOMBIE_HORSE
+            )
+
+        this.tag(AVOIDS_SCARE_SKELETON_GENE)
+            .add(
+                EntityType.SKELETON,
+                EntityType.STRAY,
+                EntityType.BOGGED,
+                EntityType.WITHER_SKELETON,
+                EntityType.SKELETON_HORSE
+            )
+
+        this.tag(AVOIDS_SCARE_SPIDER_GENE)
+            .add(
+                EntityType.SPIDER,
+                EntityType.CAVE_SPIDER
+            )
     }
 
 }