+++ CHANGELOG.md
@@ -137,3 +137,4 @@
 - GMO Recipes will have their logic printed in the tooltip of the Potion of Cell Growth in the Advanced Incubator
 - Fixed the Blood Purifier page in the Patchouli book having the Plasmid Injector instead
 - If the items given by the Keep Inventory Gene don't fit in your inventory, they're dropped at your feet instead of being tossed from your camera
+- The Scare Genes can now be given to entities